tracing: FieldInterpolator_python -> emfields.interpolator_python
